Move from_file_like class method to DakotaBase
It can be called by all subclasses. Nice! It may go away, though,
since it's likely only necessary to create a Dakota object (not a
Dakota method object) from a config file.
